+++ ios/Classes/BarcodeScanPlugin.h
@@ -1,5 +1,0 @@
-#import <Flutter/Flutter.h>
-#import "MTBBarcodeScanner.h"
-
-@interface BarcodeScanPlugin : NSObject<FlutterPlugin>
-@end